requirements.txt
discord.py==2.4.0
aiohttp==3.8.5
python-dotenv==1.0.1
pytz==2024.1
colorama==0.4.6
